% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_format.R
\name{check_bvn}
\alias{check_bvn}
\title{Check if data is a valid bivariate normal vector}
\usage{
check_bvn(bvn)
}
\arguments{
\item{bvn}{A numeric vector that is a bivariate normal vector (2 elements) or
a dataframe where each of the rows is a bivariate normal vector}
}
\description{
Check if data is a valid bivariate normal vector
}
