% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_format.R
\name{check_simplex}
\alias{check_simplex}
\title{Check if data is a valid simplex}
\usage{
check_simplex(simplex)
}
\arguments{
\item{simplex}{A numeric vector that is a 2-simplex (3 elements that sum to 1)
or a dataframe where each of the rows is a 2-simplex}
}
\description{
Check if data is a valid simplex
}
