% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_itm.R
\name{fit_itm}
\alias{fit_itm}
\title{Fit Item Response Theory Model}
\usage{
fit_itm(
  df_simplex,
  id_person,
  id_item,
  item_labels = NULL,
  iter_sampling = 500,
  iter_warmup = 500,
  n_chains = 4,
  n_cores = 1,
  adapt_delta = 0.9,
  ...
)
}
\arguments{
\item{df_simplex}{A dataframe containing the simplex data.}

\item{id_person}{A vector of person indices.}

\item{id_item}{A vector of item indices.}

\item{item_labels}{A vector of item labels. Can be long format matching id_item or
a vector of unique labels in ascending order. Default is NULL.}

\item{iter_sampling}{An integer specifying the number of sampling iterations. Default is 500.}

\item{iter_warmup}{An integer specifying the number of warmup iterations. Default is 500.}

\item{n_chains}{An integer specifying the number of Markov chains. Default is 4.}

\item{n_cores}{An integer specifying the number of cores to use. Default is 1.}

\item{adapt_delta}{A numeric value specifying the target acceptance rate. Default is 0.9.}

\item{...}{Additional arguments passed to the \code{\link[rstan:sampling]{rstan::sampling}} function.}
}
\value{
A fitted Stan model object.
}
\description{
This function fits an Item Response Theory (IRT) model using Stan.
}
\examples{
\dontrun{
df_simplex <- data.frame(matrix(runif(100), nrow=10))
id_person <- rep(1:5, each=2)
id_item <- rep(1:2, times=5)
fit <- fit_itm(df_simplex, id_person, id_item)
}
}
