% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_intervals_cumulative.R
\name{gather_values}
\alias{gather_values}
\title{Gather Values from Intervals}
\usage{
gather_values(
  lower,
  upper,
  cluster_id = NULL,
  weighted = NULL,
  step_size = NULL,
  n_samples = NULL
)
}
\arguments{
\item{lower}{A numeric vector of lower bounds of the intervals.}

\item{upper}{A numeric vector of upper bounds of the intervals.}

\item{cluster_id}{An optional vector of cluster IDs. If not provided, a vector of ones will be created.}

\item{weighted}{A logical value indicating whether the values should be weighted by interval width.}

\item{step_size}{A numeric value specifying the step size when \code{weighted} is FALSE.}

\item{n_samples}{A numeric value specifying the number of samples when \code{weighted} is TRUE.}
}
\value{
A data frame with columns \code{samples} and \code{cluster_id}.
}
\description{
This function gathers values from specified intervals, either weighted by interval width or not.
}
\details{
\itemize{
\item If \code{weighted} is TRUE, values are sampled uniformly within each interval.
\item If \code{weighted} is FALSE, values are gathered at regular intervals specified by \code{step_size}.
}
}
