% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_intervals_cumulative.R
\name{ggplot_cumulative_intervals}
\alias{ggplot_cumulative_intervals}
\title{Plot Cumulative Intervals}
\usage{
ggplot_cumulative_intervals(
  data,
  min = NULL,
  max = NULL,
  binwidth = NULL,
  facet_wrap = FALSE,
  show_quantiles = TRUE,
  ncol = 2
)
}
\arguments{
\item{data}{A data frame containing the data to be plotted.}

\item{min}{The minimum value for the x-axis.}

\item{max}{The maximum value for the x-axis.}

\item{binwidth}{The width of the bins for the histogram.}

\item{facet_wrap}{A logical value indicating whether to use facet wrapping.}

\item{show_quantiles}{A logical value indicating whether to show quantiles on the plot.}

\item{ncol}{The number of columns for facet wrapping if design matrix is not provided.}

\item{design}{An optional design matrix for facet wrapping.}
}
\value{
A ggplot2 object representing the cumulative interval plot.
}
\description{
This function creates a cumulative interval plot using ggplot2.
}
