% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_functions.R
\name{ilr}
\alias{ilr}
\title{Isometric Log-Ratio (ILR) transformation for interval responses}
\usage{
ilr(simplex)
}
\arguments{
\item{simplex}{A numeric vector that is a 2-simplex (3 elements that sum to 1) or a dataframe where each of the rows is a 2-simplex}
}
\value{
A numeric vector with 2 unbounded elements or a dataframe where each of the rows is a numeric vector with 2 unbounded elements
}
\description{
Transform 2-simplex data to the unbounded space so that they can be modeled by a bivariate-normal distribution.
This is done via a modified multivariate logit transformation that preserves the symmetry of interval responses which are conceptualized in terms of a location and a width.
}
\examples{
simplex <- data.frame(rbind(c(.1, .2, .7), c(.4, .5, .1)))
ilr(simplex)


}
