% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervalpsych-package.R
\docType{package}
\name{intervalpsych-package}
\alias{intervalpsych-package}
\alias{intervalpsych}
\title{The 'intervalpsych' package.}
\description{
The intervalpsych package provides tools for analyzing
continuous bounded interval responses in psychometrics. It implements the
Interval Consensus Model (ICM, Kloft et al., 2024), which is a consensus model
for such data.
}
\references{
Kloft, M., Siepe, B. S., & Heck, D. W. (2024).
The Interval Truth Model: A Consensus Model for Continuous Bounded Interval Responses.
\doi{doi:10.31234/osf.io/dzvw2}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://matthiaskloft.github.io/intervalpsych/}
  \item \url{https://github.com/matthiaskloft/intervalpsych}
  \item Report bugs at \url{https://github.com/matthiaskloft/intervalpsych/issues}
}

}
\author{
\strong{Maintainer}: Matthias Kloft \email{kloft.dev+intervalpsych@gmail.com} (\href{https://orcid.org/0000-0003-1845-6957}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Björn S. Siepe \email{bjoernsiepe@gmail.com} (\href{https://orcid.org/0000-0002-9558-4648}{ORCID})
}

}
\keyword{internal}
