% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_functions.R
\name{inv_ilr}
\alias{inv_ilr}
\title{Inverse Isometric Log-Ratio (ILR) transformation for interval responses}
\usage{
inv_ilr(bvn)
}
\arguments{
\item{bvn}{A numeric vector containing an unbounded interval location and width or
a dataframe where each of the rows consists of such a vector.}
}
\value{
A numeric vector containing a 2-simplex or a dataframe where each of
the rows consists of such a vector.
}
\description{
Transform unbounded data back to the 2-simplex space.
}
\examples{
simplex <- data.frame(rbind(c(0, .2), c(-2, .4)))
inv_ilr(simplex)


}
