% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_functions.R
\name{inv_slr}
\alias{inv_slr}
\title{Inverse Sum Log-Ratio (SLR) transformation for interval responses}
\usage{
inv_slr(bvn)
}
\arguments{
\item{bvn}{A numeric vector containing an unbounded interval location and width or
a dataframe where each of the rows consists of such a vector.}
}
\value{
A numeric vector containing a 2-simplex or a dataframe where each of
the rows consists of such a vector.
}
\description{
Transform unbounded data back to the 2-simplex space using the inverse sum log-ratio transformation.

The inverse SLR transformation equations are:
\deqn{x_1 = \frac{\exp(y_1)}{(\exp(y_1) + 1)(\exp(y_2) + 1)}}
\deqn{x_2 = \frac{\exp(y_2)}{\exp(y_2) + 1}}
\deqn{x_3 = \frac{1}{(\exp(y_1) + 1)(\exp(y_2) + 1)}}

where \eqn{(y_1, y_2)} are the unbounded coordinates and \eqn{(x_1, x_2, x_3)} is the resulting 2-simplex.
}
\examples{
bvn <- data.frame(rbind(c(0, .2), c(-2, .4)))
inv_slr(bvn)


}
\references{
Smithson, M., & Broomell, S. B. (2024). Compositional data analysis tutorial. Psychological Methods, 29(2), 362–378.
}
