% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_icm_stanfit.R
\name{plot.icm_stanfit}
\alias{plot.icm_stanfit}
\title{Plot Method for icm_stanfit Objects}
\usage{
\method{plot}{icm_stanfit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{icm_stanfit}.}

\item{...}{Additional arguments passed to the \code{plot_consensus} function.}
}
\value{
A plot generated by the \code{plot_consensus} function.
}
\description{
This function provides a plot method for objects of class \code{icm_stanfit}.
}
