% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_consensus.R
\name{plot_consensus}
\alias{plot_consensus}
\title{Plot ITM Consensus Intervals}
\usage{
plot_consensus(itm_stanfit, method = "median_bounds", CI = 0.95)
}
\arguments{
\item{itm_stanfit}{An object of class 'itm_stanfit' containing the Stanfit results.}

\item{method}{A character string specifying the plotting method. Options are "median_bounds" (default) or "draws_distribution".}

\item{CI}{A numeric value specifying the confidence interval for the median bounds method. Default is 0.95.}
}
\value{
A ggplot2 object representing the interval plot.
}
\description{
This function generates plots for ITM Stanfit results using either the median bounds method or the draws gradient method.
}
\details{
The function checks if the input object is of class 'itm_stanfit'. It then extracts the draws for the lower and upper bounds (T_L and T_U) from the Stanfit object.

If the method is "median_bounds", the function calculates the median of the lower and upper bounds and generates an interval plot using these medians.

If the method is "draws_distribution", the function checks the number of draws and warns if it is less than 1000. It then computes a consensus distribution for each item and generates a gradient interval plot.
}
