% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_functions.R
\name{slr}
\alias{slr}
\title{Sum Log-Ratio (SLR) transformation for interval responses}
\usage{
slr(simplex)
}
\arguments{
\item{simplex}{A numeric vector that is a 2-simplex (3 elements that sum to 1) or a dataframe where each of the rows is a 2-simplex}
}
\value{
A numeric vector with 2 unbounded elements or a dataframe where each of the rows is a numeric vector with 2 unbounded elements
}
\description{
Transform 2-simplex data to the unbounded space using sum log-ratio transformation.
This transformation provides an alternative to the ILR transformation for modeling interval data.

The SLR transformation equations are:
\deqn{y_1 = \log\left(\frac{x_1}{x_3}\right)}
\deqn{y_2 = \log\left(\frac{x_2}{x_1 + x_3}\right)}

where \eqn{(x_1, x_2, x_3)} is a 2-simplex and \eqn{(y_1, y_2)} are the transformed coordinates.
}
\examples{
simplex <- data.frame(rbind(c(.1, .2, .7), c(.4, .5, .1)))
slr(simplex)


}
\references{
Smithson, M., & Broomell, S. B. (2024). Compositional data analysis tutorial. Psychological Methods, 29(2), 362–378.
}
