% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_functions.R
\name{splx_to_itvl}
\alias{splx_to_itvl}
\title{Convert from simplex to interval bounds}
\usage{
splx_to_itvl(simplex, min = NULL, max = NULL)
}
\arguments{
\item{simplex}{Data in the simplex/compositional format.}

\item{min}{Minimum of the original response scale.}

\item{max}{Maximum of the original response scale.}
}
\description{
Convert from simplex/compostional format to interval bound format
}
\examples{
responses <- data.frame(rbind(c(.1,.5,.4), c(.3,.4,.3)))
splx_to_itvl(responses, min = 0, max = 1)


}
