% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_icm_stanfit.R
\name{summary.icm_stanfit}
\alias{summary.icm_stanfit}
\title{Summarize Stanfit Object}
\usage{
\method{summary}{icm_stanfit}(object, ...)
}
\arguments{
\item{object}{An object of class \code{icm_stanfit}.}

\item{...}{Additional arguments (currently not used).}
}
\value{
A summary of the \code{icm_stanfit} object.
}
\description{
This function provides a summary for an object of class \code{icm_stanfit}.
}
\examples{
\dontrun{
  # Assuming `fit` is an object of class `icm_stanfit`
  summary(fit)
}

}
