% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_itm.R
\name{theme_itm}
\alias{theme_itm}
\title{Custom ggplot2 Theme for intervalpsych}
\usage{
theme_itm(hide_axis_text_y = FALSE, base_size = 12)
}
\arguments{
\item{hide_axis_text_y}{Logical. If TRUE, the y-axis text and ticks will be hidden. Default is FALSE.}

\item{base_size}{Numeric. Base font size for the theme. Default is 12.}
}
\value{
A ggplot2 theme object.
}
\description{
This function creates a custom ggplot2 theme for intervalpsych visualizations.
}
