% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_consensus.R
\name{extract_consensus}
\alias{extract_consensus}
\title{Extract Consensus intervals from ITM Stan Fit Object}
\usage{
extract_consensus(itm_stanfit, print_summary = TRUE)
}
\arguments{
\item{itm_stanfit}{An object of class \code{itm_stanfit} containing the fitted Stan model.}

\item{print_summary}{A logical value indicating whether to print a summary of the extracted parameters. Default is \code{TRUE}.}
}
\value{
A list containing:
\item{df_rvar}{A data frame with extracted posterior samples as random variables.}
\item{summary}{A summary data frame with median and credible intervals for the extracted parameters.}
\item{item_labels}{A character vector of item labels.}
}
\description{
This function extracts parameter estimates for the consensus intervals from a
fitted Interval Truth Model Stan fit object of class \code{itm_stanfit}.
}
\details{
This function extracts parameter estimates for the consensus intervals from a
fitted Interval Truth Model Stan fit object of class \code{itm_stanfit}.
}
