% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_functions.R
\name{itvl_to_splx}
\alias{itvl_to_splx}
\title{Convert from interval bounds to simplex}
\usage{
itvl_to_splx(interval_bounds, min = NULL, max = NULL)
}
\arguments{
\item{interval_bounds}{Data in the interval bounds format.}

\item{min}{Minimum of the original response scale.}

\item{max}{Maximum of the original response scale.}
}
\description{
Convert interval responses from interval bounds format to compostional/simplex format
}
\examples{
interval_responses <- data.frame(rbind(c(.1,.5), c(.4,.7)))
itvl_to_splx(interval_responses, min = 0, max = 1)

}
