% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_icm_stanfit.R
\name{plot.icm_stanfit}
\alias{plot.icm_stanfit}
\title{Plot Method for icm_stanfit Objects}
\usage{
\method{plot}{icm_stanfit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{icm_stanfit}.}

\item{...}{Additional arguments passed to the \code{plot_consensus} function.}
}
\value{
A plot generated by the \code{plot_consensus} function.
}
\description{
This function provides a plot method for objects of class \code{icm_stanfit}.
}
\examples{
\dontrun{
# Fit ICM model first
df_simplex <- data.frame(matrix(runif(100), nrow=10))
id_person <- rep(1:5, each=2)
id_item <- rep(1:2, times=5)
fit <- fit_icm(df_simplex, id_person, id_item)

# Plot using S3 method
plot(fit)

# Plot with specific method and parameters
plot(fit, method = "draws_distribution", CI = 0.90)
}

}
