% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_zeros.R
\name{remove_zeros}
\alias{remove_zeros}
\title{Remove Zeros from Simplex}
\usage{
remove_zeros(simplex, method = "simple_replacement", padding = 0.01)
}
\arguments{
\item{simplex}{A numeric matrix, data frame, or vector representing the simplex. Each row should sum to one and all elements should be between 0 and 1.}

\item{method}{A character string specifying the method to remove zeros. Currently, only "simple_replacement" is supported. Default is "simple_replacement".}

\item{padding}{A numeric value to add to each element of the simplex when using the "simple_replacement" method. Default is 0.01.}
}
\value{
A numeric matrix with the same dimensions as the input \code{simplex}, with zeros removed according to the specified method.
}
\description{
This function removes zeros from a simplex (a matrix or vector where each row sums to one and all elements are between 0 and 1) using a specified method.
}
\examples{
# Example usage:
simplex <- matrix(c(0.2, 0.3, 0.5, 0, 0.5, 0.5), nrow = 2, byrow = TRUE)
remove_zeros(simplex)

}
